Track: Move projection code to TemplateTrack
Projected coordinates from TrackPoint are used by only TemplateTrack,
for actual drawing. This change makes TemplateTrack store the data
as close possible to what is needed for fast drawing, e.g. using
QPainterPath for track segments. So this includes a significant
rewrite of the painting in TemplateTrack.

Attention: This change breaks the updating of TemplateTracks during
recording. This needs to be solved in a subsequent change.